feat: add the macro branch to test the sqrt
