+++ src/secp256k1/fp.rs
@@ -300,25 +300,5 @@ mod test {
     crate::field_testing_suite!(Fp, "bits");
     crate::field_testing_suite!(Fp, "serialization_check");
     crate::field_testing_suite!(Fp, "constants", MODULUS_STR);
-
-    use rand_core::OsRng;
-
-    #[test]
-    fn test_sqrt() {
-        // NB: TWO_INV is standing in as a "random" field element
-        let v = (Fp::TWO_INV).square().sqrt().unwrap();
-        assert!(v == Fp::TWO_INV || (-v) == Fp::TWO_INV);
-
-        for _ in 0..10000 {
-            let a = Fp::random(OsRng);
-            let mut b = a;
-            b = b.square();
-
-            let b = b.sqrt().unwrap();
-            let mut negb = b;
-            negb = negb.neg();
-
-            assert!(a == b || a == negb);
-        }
-    }
+    crate::field_testing_suite!(Fp, "sqrt");
 }
